Refactoring files structure
