+++ src/socket.cpp
@@ -1,7 +1,8 @@
-#include "socket.h"
+#include "../include/socket.h"
 #include <string.h>
 #include <errno.h>
 #include <fcntl.h>
+#include <iostream>
 
 
 Socket::Socket() : m_sock (-1) // -1 means INVALID SOCKET
