+++ src/whispr_server.cpp
@@ -1,4 +1,4 @@
-#include "../include/server_socket.h"
+#include "../include/socket.h"
 
 int main() {
   Socket();
